+++ loguru/_datetime.py
@@ -2,7 +2,7 @@ import re
 from calendar import day_abbr, day_name, month_abbr, month_name
 from datetime import datetime as datetime_
 from datetime import timedelta, timezone
-from time import localtime
+from time import localtime, strftime
 
 tokens = r"H{1,2}|h{1,2}|m{1,2}|s{1,2}|S{1,6}|YYYY|YY|M{1,4}|D{1,4}|Z{1,2}|zz|A|X|x|E|Q|dddd|ddd|d"
 
@@ -74,6 +74,16 @@ class datetime(datetime_):
 
 def now():
     now = datetime.now()
-    local = localtime(now.timestamp())
-    tzinfo = timezone(timedelta(seconds=local.tm_gmtoff), local.tm_zone)
+    timestamp = now.timestamp()
+    local = localtime(timestamp)
+
+    try:
+        seconds = local.tm_gmtoff
+        zone = local.tm_zone
+    except AttributeError:
+        offset = datetime.fromtimestamp(timestamp) - datetime.utcfromtimestamp(timestamp)
+        seconds = offset.total_seconds()
+        zone = strftime("%Z")
+
+    tzinfo = timezone(timedelta(seconds=seconds), zone)
     return now.replace(tzinfo=tzinfo)